feat: 커뮤니티 상세 페이지 로직 분리
