+++ apps/client/src/widgets/community/components/community-body/community-body.css.ts
@@ -1,10 +1,5 @@
 import { style } from '@vanilla-extract/css';
 
-export const container = style({
-  display: 'flex',
-  flexDirection: 'column',
-});
-
 export const bottomFloating = style({
   position: 'fixed',
   bottom: '2.4rem',
